server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
cryptography==42.0.5

